pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "purview-governance"
version = "1.0.0"
description = "Multi-tenant GCC-compliant Purview & Compliance Manager AI governance solution"
requires-python = ">=3.11"
dependencies = [
    "msal>=1.28.0",
    "requests>=2.32.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "click>=8.1.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    # Function App dependencies needed to import shared modules in tests
    "azure-functions>=1.18.0",
    "azure-data-tables>=12.5.0",
    "azure-identity>=1.16.0",
    "jsonschema>=4.23.0",
]

[project.scripts]
purview-collect = "collector.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
